Split database/API and remove database location in the package.json.
Sorry this PR introduce a couple of changes:
* Split the the hosting in multiple files for readability
* Remove the database location in the package.json file
* You need to give path to the database file when you use it
* No fallback if the database is missing, instead you will get an error
